+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [macOS] Update the design to use the transparent toolbar and move the inspectors to the native left sidebar 
 - [macOS] Fix an issue with navigation from the metrics screen not working correctly
 - [macOS] Move the button to switch from the horizontal to vertical layout to the details view 
+- [macOS] Fix an issue with text view not scrolling to the selected match when opened from the full-body console search
 
 ## Pulse 3.4.1